# OpenAPI.yaml
...
openapi: 3.1.0
info:
  title: InferaDB Management API
  description: |
    Control Plane API for InferaDB providing self-service user authentication, organization management, and vault access control.

    ## Authentication

    The Management API uses two primary authentication methods:

    - **Session Cookie**: For web dashboard and interactive sessions. Set via `infera_session` cookie after login.
    - **JWT Bearer Token**: For programmatic access via client assertions (OAuth 2.0 JWT Bearer flow).

    ## Multi-Tenancy

    All resources are scoped to organizations. Users can belong to multiple organizations with different roles:

    - **Owner**: Full control, can transfer ownership
    - **Admin**: Manage members, vaults, clients, and teams
    - **Member**: Basic access, can view organization resources

    ## Rate Limiting

    API requests are rate-limited by IP address and user:

    - Login: 100 attempts/hour per IP
    - Registration: 5 per day per IP
    - Email verification: 5 tokens/hour
    - Password reset: 3 tokens/hour

    Rate limit headers are included in responses:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

  version: 1.0.0
  contact:
    name: InferaDB Support
    url: https://inferadb.com/docs
    email: support@inferadb.com
  license:
    name: Business Source License 1.1
...
servers:
  - url: http://localhost:3000
...
paths:
...
  /v1/users/me:
    get:
      tags: [Users]
      summary: Get user profile
      operationId: getUserProfile
      security:
        - SessionCookie: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"

    patch:
      tags: [Users]
      summary: Update user profile
      operationId: updateUserProfile
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Alice Johnson
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

    delete:
      tags: [Users]
      summary: Delete user account
      description: Soft delete user account (can be recovered within 30 days)
      operationId: deleteUser
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Account deleted
        "401":
          $ref: "#/components/responses/Unauthorized"

  # Email Management Endpoints
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: infera_session
      description: Session cookie obtained from /v1/auth/login or /v1/auth/register
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: invalid_input
        message:
          type: string
          example: Email address is invalid
        details:
          type: object
          description: Additional error context
    User:
      type: object
      required: [id, name, created_at]
      properties:
        id:
          type: string
          format: snowflake
          example: "123456789012345678"
        name:
          type: string
          example: Alice Smith
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
...